Agregar sede a los listados de canchas, precios enteros y scripts de setup
- Asociar Court con Location (faltaba en association.js) e incluir la sede
  en /seeCourts y /canchas/verCanchas, para que el listado pueda mostrarla.
- Pasar los precios de DECIMAL a INTEGER en Cancha, Servicio y Reserva: el
  negocio no usa centavos. De paso arregla el overflow de priceService, que
  con DECIMAL(5,2) fallaba con cualquier monto de $1000 o mas.
- Exigir precio entero en la validacion de canchas y servicios.
- Aclarar el 409 de deleteCourt: los horarios son franjas configuradas, no
  reservas. Ahora dice cuantos hay y las dos formas de resolverlo. Idem en
  deleteHorary con la cantidad de reservas.
- Agregar .catch al sync() para que un fallo de conexion muestre un mensaje
  claro en vez de terminar en unhandled rejection.
- Agregar scripts npm: crear-admin, necesario porque el registro publico
  fuerza typeUser CLIENTE y sin un ADMIN no se puede entrar al panel; y
  revisar-db, diagnostico de solo lectura del esquema.

Nota: el cambio de precios requiere migrar las columnas existentes, porque
sync({ alter: false }) no modifica tablas ya creadas:
  ALTER TABLE canchas   MODIFY hourlyPrice  INT NOT NULL;
  ALTER TABLE servicios MODIFY priceService INT NOT NULL;
  ALTER TABLE reservas  MODIFY totalAmount  INT NOT NULL;

+++ src/models/Reserva.js
@@ -19,7 +19,7 @@ export const Reserve = sequelize.define("Reserva", {
     },
     
     totalAmount: {
-      type: DataTypes.DECIMAL(7, 2),
+      type: DataTypes.INTEGER,
       allowNull: false
     },
